Video 14

+++ admin/ajax/features_facilities.php
@@ -14,9 +14,9 @@ if (isset($_POST['add_feature'])) {
     echo $res;
 }
 
-
 if (isset($_POST['get_features'])) {
     $res = selectAll('features');
+    $i = 1;
 
     while ($row = mysqli_fetch_assoc($res)) {
         echo <<<data
@@ -38,9 +38,15 @@ if (isset($_POST['rem_feature'])) {
     $frm_data = filteration($_POST);
     $values = [$frm_data['rem_feature']];
 
-    $q = "DELETE FROM `features` WHERE `id`=?";
+    $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?', [$frm_data['rem_feature']], 'i');
+
+    if (mysqli_num_rows($check_q)==0) {
+        $q = "DELETE FROM `features` WHERE `id`=?";
     $res = delete($q,$values,'i');
     echo $res;
+    } else {
+        echo 'room_added';
+    } 
 }
 
 if (isset($_POST['add_facility'])) {
@@ -48,59 +54,61 @@ if (isset($_POST['add_facility'])) {
 
     $img_r = uploadSVGImage($_FILES['icon'],FACILITIES_FOLDER);
 
-    if ($img_r === 'inv_img') {
+    if ($img_r === `inv_img`) {
         echo $img_r;
     } 
-    else if ($img_r === 'inv_size') {
+    else if ($img_r === `inv_size`) {
         echo $img_r;
     } 
-    else if ($img_r === 'upd_failed') {
+    else if ($img_r === `upd_failed`) {
         echo $img_r;
     } else {
-        $q = "INSERT INTO `facilities` (`icon`, `name`, `description`) VALUES (?,?)";
-        $values = [$img_r, $frm_data['name'], $frm_data['desc']];
+        $q = "INSERT INTO `facilities` (`icon`, `name`, `description`) VALUES (?,?,?)";
+        $values = [$img_r, $frm_data['name'], $frm_data['desc']]; 
         $res = insert($q, $values, 'sss');
         echo $res;
-        }
     }
+}
 
-    if (isset($_POST['get_facilities'])) {
-        $res = selectAll('facilities');
-        $i=1;
-        $path = FACILITIES_IMG_PATH;
-    
-        while ($row = mysqli_fetch_assoc($res)) {
-            echo <<<data
-            <tr>
-                <td>$i</td>
-                <td><img src="$path$row[icon]" width="100px"></td>
-                <td>$row[name]</td>
-                <td>$row[description]</td>
-                <td>
-                    <button type="button" onclick="rem_feature($row[id])" class="btn btn-danger btn-sm shadow-none">
+if (isset($_POST['get_facilities'])) {
+    $res = selectAll('facilities');
+    $i=1;
+    $path = FACILITIES_IMG_PATH;
+
+    while ($row = mysqli_fetch_assoc($res)) {
+        echo <<<data
+        <tr>
+            <td>$i</td>
+            <td><img src="$path$row[icon]" width="100px"></td>
+            <td>$row[name]</td>
+            <td>$row[description]</td>
+            <td>
+                <button type="button" onclick="rem_facility($row[id])" class="btn btn-danger btn-sm shadow-none">
                     <i class="fa fa-trash"></i> Delete
-            </button>
-                </td>
-            </tr>
-        data;
-        $i++;
-        }
+                </button>        
+            </td>
+        </tr>
+    data;
+    $i++;
     }
+}
 
-    if (isset($_POST['rem_facility'])) {
-        $frm_data = filteration($_POST);
-        $values = [$frm_data['rem_facility']];
-    
-        $q = "DELETE FROM `facilities` WHERE `id`=?";
-        $res = select($q,$values,'i');
-        $img = mysqli_fetch_assoc($res);
+if (isset($_POST['rem_facility'])) {
+    $frm_data = filteration($_POST);
+    $values = [$frm_data['rem_facility']];
 
-        if(deleteImage($img['icon'], FACILITIES_FOLDER)){
-            $q = "DELETE FROM  `facilities` WHERE `id`=?";
-            $res = delete($q,$values,'i');
-            echo $res;
-        }
+    $q = "DELETE FROM `facilities` WHERE `id`=?";
+    $res = select($q,$values,'i');
+    $img = mysqli_fetch_assoc($res);
+
+    if(deleteImage($img['icon'], FACILITIES_FOLDER)){
+        $q = "DELETE FROM  `facilities` WHERE `id`=?";
+        $res = delete($q,$values,'i');
         echo $res;
     }
+    else {
+        echo 0;
+    }
+}
 
 ?>
